Use queue of completed wave headers.
In the unlikely event that several wave headers are completed
before the wait, lastWaveHdr would have been overwritten before
getNextFragment() could call waveInUnprepareHeader(). This change
ensures waveInUnprepareHeader() will always be called and the last
complete header will be used for data. This is an infinite loop
if header processing can't ever catch up, but that should be
extremely unlikely.