Add 32 bpp truecolor support.
This corresponds to the 32 bpp patch from my old Windows port.
It is integrated in a way which allows use of either 8 bpp or
32 bpp depending on the --truecolor command line flag.

The 32 bpp mode uses a large 32 bits * 65536 (256 KB) lookup table
for converting from the two dimensional internal colour space to
RGB. The table is constructed in setupPalette() in ui.cc and
passed via screen->setPalette(). It uses the most common
arrangement, ignoring SDL_PixelFormat.