Add support for compiling a Winamp plugin via --enable-winamp.
This integrates relevant code from the my old Windows port, which
ran as a Winamp plugin. Because Winamp only supplies 576 8-bit
samples, visualization quality is degraded. The configuration file
is stored at the standard location, and not the Winamp plugin
location.

Building is done via a hack, setting LDFLAGS instead of using
libtool. Adding libtool support seems too complicated and
pointless. It will build synaesthesia.exe which is in fact a DLL.
It needs to be renamed to a file of the form vis_*.dll and copied
into the Winamp visualization plugin directory. SDL.dll needs to
be in the path. Header files vis.h and wa_ipc.h from Winamp SDK
are required but not included.

Loading a GPL licensed plugin into a GPL-incompatible program
violates the GPL according to the GPL FAQ. I don't see a problem
with distributing this code, but I will not distribute binaries.