Add AM_PATH_SDL fallback using sdl-config to find SDL 1.2
Windows SDL fails to ship the pkg-config sdl.pc file.
It does ship sdl-config, so this will find it, but unfortunately
builds will still fail because this won't support static linking.
That build failure is easier to deal with at least. Build SDL
from source to get sdl.pc and avoid this mess.