Add frames per second setting for PortAudio and SDL audio
This adds the --fps command line argument and fps configuration
file setting for visual frames per second. A new frame is produced
when Synaesthsia is finished with the previous frame and new sound
samples arrive. The sound samples arive in groups, and frame rate
is controlled by changing the size of those groups at startup.